# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NewPeace BookStore API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/auth/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "訂單紀錄",
                "operationId": "3d52aa29e9a87a39614334e40d3c5050",
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1,
                                    "data": [
                                        {
                                            "id": 1,
                                            "user_id": 1,
                                            "status": 1,
                                            "total_price": 100,
                                            "created_at": null,
                                            "updated_at": null
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/order/{oid}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "指定訂單紀錄",
                "operationId": "da3603393acb2be073300b7780dca010",
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1,
                                    "data": {
                                        "id": 1,
                                        "user_id": 1,
                                        "status": 1,
                                        "total_price": 100,
                                        "created_at": null,
                                        "updated_at": null
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "失敗(訂單不存在)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "order not found"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "失敗(訂單不屬於此會員)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "order not found"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{id}/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "會員訂單紀錄",
                "operationId": "9d7b1a365d25e8b27db78493c077717e",
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1,
                                    "data": [
                                        {
                                            "id": 1,
                                            "user_id": 1,
                                            "status": 1,
                                            "total_price": 100,
                                            "created_at": null,
                                            "updated_at": null
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "失敗(會員不存在)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "user not found"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{id}/order/{oid}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "會員指定訂單紀錄",
                "operationId": "c1fb86604b998588ddb0a25d88888162",
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1,
                                    "data": {
                                        "id": 1,
                                        "user_id": 1,
                                        "status": 1,
                                        "total_price": 100,
                                        "created_at": null,
                                        "updated_at": null
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "失敗(訂單不存在)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "order not found"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "失敗(訂單不屬於此會員)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "order not belongs this user"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/marchant/manage": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "商家取得訂單",
                "operationId": "5d397dec4fffbc23978e84b9ff20981a",
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1,
                                    "todo": [
                                        {
                                            "id": 1,
                                            "user_id": 1,
                                            "status": 1,
                                            "total_price": 100,
                                            "created_at": null,
                                            "updated_at": null
                                        }
                                    ],
                                    "done": [
                                        {
                                            "id": 1,
                                            "user_id": 1,
                                            "status": 1,
                                            "total_price": 100,
                                            "created_at": null,
                                            "updated_at": null
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/marchant/manage/{oid}/payment/complete": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "商家手動標示訂單已付款(賣家或管理員可操作)",
                "operationId": "66eea4671dc0f26e66c66fad007b3fce",
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/marchant/manage/{oid}/complete": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "商家標示訂單已完成(賣家或管理員可操作)",
                "operationId": "a98651481ca52cb522609bdd1e936adb",
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}